Move DKG test loop chain function stub implementations from v1 chain to v2 chain handle
Since `beacon.Initialize` uses the new
v2 beacon chain implementation to run the DKG
test loop, we need to port all stub
implementations of the chain functions used
by the DKG loop from the legacy v1 chain handle
to the new v2 beacon chain implementation.
